BB-190: Split identifiers components into multiple files
